backend/requirements.txt
asgiref==3.3.1
Django==3.1.7
django-extensions==3.1.1
django-rest==0.8.6
djangorestframework==3.12.2
psycopg2-binary==2.8.6
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
